refactor: extract mint components and remove unused styled components
Move MintLayout, Panel, StatusBox, ResponseBlock from wrappers.tsx into
new mint.tsx module. Remove unused components: PrimaryButton, FormWrapper,
FormInput, FormTextArea, FormValidation, Blockquote, Label, OrderedList,
Logotype, Article, Credits, FooterAside. Clean up globalstyle.tsx.
Update imports in mint/page.tsx and MintActions.tsx.
